4ano/RAS-Project/exam-assessment/requirements.txt
Django >= 5.0
djangorestframework >= 3.14.0
django-extensions >= 3.2.3
mysqlclient >= 2.2.1
spacy >= 3.7.2
